+++ nes-nautilus/include/Nautilus/Interface/BufferRef/OutputFormatterBufferRef.hpp
@@ -66,13 +66,13 @@ public:
         const RecordBuffer& recordBuffer,
         nautilus::val<uint64_t>& recordIndex) const override;
 
-    nautilus::val<size_t> writeRecord(
+    nautilus::val<uint64_t> writeRecord(
         nautilus::val<uint64_t>& bytesWritten,
         const RecordBuffer& recordBuffer,
         const Record& rec,
         const nautilus::val<AbstractBufferProvider*>& bufferProvider) const override;
 
-    nautilus::val<size_t> writeRecordSafely(
+    nautilus::val<uint64_t> writeRecordSafely(
         nautilus::val<uint64_t>& bytesWritten,
         const RecordBuffer& recordBuffer,
         const Record& rec,
